Refactor core renderer: extract SeatDataManager, ViewportController, and optimize dependencies.
